Lift the file-local type walk out of DirectInstantiation
Pure refactor, no behaviour change: `PrivateTypeDeclarationCollector` becomes
`FileLocalTypeCollector` in its own file so a second rule can use it.

`ConcreteTypeUsage` is that second rule. The two fire on the same seam from
opposite ends — the construction site and the declared type — and they have twice
before been found holding separate copies of the same vocabulary:
`ServiceTypeSuffix` and then `MockTypeName`, each time after the rule that lacked
it reported something its twin already knew was not a finding. Keeping this walk
in a file of its own rather than importing it from one visitor into the other is
the arrangement that stops a third.
